+++ src/test/scala/scotty/simulator/DaggerGateSpec.scala
@@ -10,10 +10,10 @@ class DaggerGateSpec extends FlatSpec with TestHelpers {
     val gate = (index: Int) => DefGate(Array(
       Array(Complex(1), Complex(0)),
       Array(Complex(0), Complex(0, -1))
-    ).toDouble, index)
+    ).toFloat, index)
 
     assert(Dagger(gate(0)).matrix(sim).toList.map(_.toList) == List(
       List(Complex(1), Complex(0)),
-      List(Complex(0), Complex(0, 1))).toDouble)
+      List(Complex(0), Complex(0, 1))).toFloat)
   }
 }
